+++ src/core/error_handling.c
@@ -1,6 +1,7 @@
 // clang-format off
 #define PY_SSIZE_T_CLEAN
 #include "Python.h"
+#include "python_unexposed.h"
 // clang-format on
 #include "error_handling.h"
 #include "jslib.h"